+++ droidlet/tools/crowdsourcing/vision_annotation_task/server_files/extra_refs/archive/voxel_wrapper_q1.html
@@ -1,0 +1,41 @@
+<html>
+    <head>
+        <link	
+            rel="stylesheet"	
+            href="https://stackpath.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"	
+        />	
+    </head>
+    <body>
+        <div id="voxel_painter"></div>
+        <div id="buttons">
+            <button type="button" class="center btn btn-secondary" id="reset">Reset to Starting State</button>
+            <button type="button" class="center btn btn-danger" id="null-result">The Object is Not Present</button>
+            <button type="button" class="center btn btn-success" id="finish-annotation">Finish Annotation</button>
+            <div id="response" class="center"></div>
+        </div>
+        <input type="hidden" id="markedBlocks" name="markedBlocks" value="">
+    </body>
+    <script type='module' id="voxelPainter" src='voxel_painter_q1.mjs'></script>
+    <script>
+        document.getElementById("finish-annotation").addEventListener("click", function() {
+            let output_obj = {q1_output: document.getElementById("markedBlocks").value};
+            window.parent.postMessage(JSON.stringify({ msg: output_obj }), "*");
+            document.getElementById("finish-annotation").disabled = true;
+            document.getElementById("null-result").disabled = true;
+            document.getElementById("response").innerHTML = " - Annotation Complete";
+        });
+        document.getElementById("null-result").addEventListener("click", function() {
+            let output_obj = {q1_output: "null"};  // parser expects a string
+            window.parent.postMessage(JSON.stringify({ msg: output_obj }), "*");
+            document.getElementById("finish-annotation").disabled = true;
+            document.getElementById("null-result").disabled = true;
+            document.getElementById("response").innerHTML = " - Annotation Complete";
+        });
+        document.getElementById("reset").addEventListener("click", function() {
+            document.getElementById("finish-annotation").disabled = false;
+            document.getElementById("null-result").disabled = false;
+            location.reload();
+        });
+    </script>
+    <link rel="stylesheet" href="voxel_wrapper.css">
+</html>
