+++ droidlet/tools/crowdsourcing/vision_annotation_task/server_files/extra_refs/utils.js
@@ -25,7 +25,7 @@ function recordClick(ele) {
   }
   clickedElements.push({id: ele, timestamp: Date.now()});
   document.getElementById("clickedElements").value = JSON.stringify(clickedElements);
-  console.log("Clicked elements array: " + JSON.stringify(clickedElements));
+  //console.log("Clicked elements array: " + JSON.stringify(clickedElements));
 }
 recordClick("start");
