+++ droidlet/tools/crowdsourcing/vision_annotation_task/server_files/vision_annotation_pilot.html
@@ -93,7 +93,7 @@
   <hr>
   
   <div class="row">
-    <h2>Question #1 - mark the blocks in the blue tower</h2>
+    <h2>Question #1 - mark the blocks between the user and the assistant</h2>
     <div class="col-xs-6">
       <h3 class="center">Original Scene</h3>
       <p class="center">
@@ -109,27 +109,6 @@
       </p>
     </div>
     <iframe style="width:100%; height:500px; margin: 0 auto;" id="myiframe" src="voxel_wrapper_q1.html" scrolling="no"></iframe>
-  </div>
-
-  <hr>
-  
-  <div class="row">
-    <h2>Question #2 - mark all the air blocks between the two pillars</h2>
-    <div class="col-xs-6">
-      <h3 class="center">Original Scene</h3>
-      <p class="center">
-        Perform annotation in the other window
-      </p>
-    </div>
-    <div class="col-xs-6">
-      <h3 class="center">Annotation Window</h3>
-      <p class="center">
-        Click in window to enter<br>
-        Use arrow keys to rotate view<br>
-        Press 'z' to undo an action
-      </p>
-    </div>
-    <iframe style="width:100%; height:500px; margin: 0 auto;" id="myiframe" src="voxel_wrapper_q2.html" scrolling="no"></iframe>
   </div>
 
   <hr>
@@ -151,27 +130,6 @@
       </p>
     </div>
     <iframe style="width:100%; height:500px; margin: 0 auto;" id="myiframe" src="voxel_wrapper_q3.html" scrolling="no"></iframe>
-  </div>
-
-  <hr>
-  
-  <div class="row">
-    <h2>Question #4 - unmark the blocks in the middle cube</h2>
-    <div class="col-xs-6">
-      <h3 class="center">Original Scene</h3>
-      <p class="center">
-        Perform annotation in the other window
-      </p>
-    </div>
-    <div class="col-xs-6">
-      <h3 class="center">Annotation Window</h3>
-      <p class="center">
-        Click in window to enter<br>
-        Use arrow keys to rotate view<br>
-        Press 'z' to undo an action
-      </p>
-    </div>
-    <iframe style="width:100%; height:500px; margin: 0 auto;" id="myiframe" src="voxel_wrapper_q4.html" scrolling="no"></iframe>
   </div>
 
   <hr>
